progresstracker cleanup
