+++ Source/ItsSorceryFramework/ItsSorceryFramework.csproj
@@ -65,6 +65,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ProgressLevelOption.cs" />
+    <Compile Include="ProgressLevelModulo.cs" />
     <Compile Include="CompAbilityEffect_AddEXP.cs" />
     <Compile Include="Hediff_ProgressLevel.cs" />
     <Compile Include="CompAbilityEffect_ForceLevelUp.cs" />
